Move the temporary directory into the root, and allow it's customization too.
This fixes a situation where you mount the Docker socket into a Docker image so
that an "isolated" (or redeployable) CI within a Docker image can't mount and
read the stuff it's mounting because it's created on the instance and not on the
system itself.  Now you can mount a dir within the instance that is shared and
set that as the temporary directory.